improved pose_detection pipeline

+++ vid/patient_vid_crop.py
@@ -13,7 +13,7 @@ def gen_cropped_frames(video_path, coords_path, save_path):
     coords = open(coords_path).readlines()
     if cap.get(cv2.cv.CV_CAP_PROP_FPS) < 29:
         append_frame = int(30/(30-cap.get(cv2.cv.CV_CAP_PROP_FPS)))
-        for i in xrange(cap.get(cv2.cv.CV_CAP_PROP_FRAME_COUNT)/append_frame):
+        for i in xrange(int(cap.get(cv2.cv.CV_CAP_PROP_FRAME_COUNT)/append_frame)):
             coords.insert(i*append_frame, coords[i*append_frame])
 
     vid = my_video_capture(video_path, frame_rate=30)
